consolidate release-notes scripts and integrate with versioning scfipts, update + fix workflow for same
