+++ .github/workflows/auto-freeze-version.yml
@@ -44,7 +44,7 @@ jobs:
 
           echo "version=$VERSION" >> $GITHUB_OUTPUT
           echo "new_dev_version=$NEW_DEV_VERSION" >> $GITHUB_OUTPUT
-          
+
           echo "Freezing version: $VERSION"
           echo "Next development version: $NEW_DEV_VERSION"
 
@@ -66,7 +66,7 @@ jobs:
           # Set non-interactive mode for the version manager
           export CI=true
           export BATCH_MODE=true
-          
+
           # Run the automated version freeze
           NEW_VERSION="$NEW_VERSION" node scripts/versioning/version-manager.js
 
@@ -87,7 +87,7 @@ jobs:
           git commit -m "docs: freeze ${{ steps.version-info.outputs.version }} and begin ${{ steps.version-info.outputs.new_dev_version }} development
 
           Automated version freeze triggered by cosmos/evm release ${{ steps.version-info.outputs.version }}
-          
+
           Changes:
           - Frozen documentation snapshot at ${{ steps.version-info.outputs.version }}/
           - Created Google Sheets tab for version-specific EIP data
@@ -115,7 +115,7 @@ jobs:
           ## Review Checklist
 
           - [ ] Verify frozen documentation renders correctly
-          - [ ] Check Google Sheets tab was created successfully  
+          - [ ] Check Google Sheets tab was created successfully
           - [ ] Confirm navigation includes new version
           - [ ] Test EIP compatibility table shows version-specific data
