+++ scripts/versioning/package.json
@@ -6,7 +6,7 @@
   "private": true,
   "scripts": {
     "freeze": "node version-manager.js",
-    "release-notes": "node release-notes.js",
+    "changelogs": "node manage-changelogs.js",
     "sheets": "node sheets-manager.js",
     "test": "node test-versioning.js"
   },
